Refactor main.rb, update README
